Split health-check mod-requirements UI into smaller modules
Break ModRequirementsContent.tsx into a ListingRow, DetailView, a
modToFileData util, an isModHidden entries helper, and move the shared
actions hook up to hooks/, mirroring the file-requirements layout. Also
move the detail card border into each DetailView so the mod disclaimer
can sit outside it, and route the detail title/disclaimer text through
translations.
