+++ src/renderer/src/extensions/health_check/views/HealthCheckDetailPage.tsx
@@ -67,9 +67,7 @@ function HealthCheckDetailPage({ api, content, entry, onBack }: IHealthCheckDeta
             </Button>
           </div>
 
-          <div className="rounded-lg border border-stroke-weak">
-            <DetailView api={api} entry={entry} onBack={onBack} />
-          </div>
+          <DetailView api={api} entry={entry} onBack={onBack} />
 
           <PremiumBanner />
         </div>
