+++ src/renderer/src/extensions/health_check/views/content/ModRequirementsContent.tsx
@@ -1,254 +1,15 @@
-import { mdiCheck, mdiDownload, mdiOpenInNew } from "@mdi/js";
-import { unknownToError } from "@vortex/shared";
-import React, { useCallback, useEffect, useMemo } from "react";
-import { Trans, useTranslation } from "react-i18next";
-import { useSelector } from "react-redux";
-
-import type { IFileRequirementData } from "@/extensions/health_check/utils/fileRequirements/cardHelpers";
-import { getModFilesWithCache } from "@/extensions/health_check/utils/modRequirements/modFiles";
 import { onDownloadRequirement } from "@/extensions/health_check/utils/modRequirements/onDownloadRequirement";
-import { log } from "@/logging";
 import type { IExtensionApi } from "@/types/IExtensionContext";
 import type { IState } from "@/types/IState";
-import { Button } from "@/ui/components/button/Button";
-import { PremiumBadge } from "@/ui/components/premium_badge/PremiumBadge";
-import { Typography } from "@/ui/components/typography/Typography";
-import { TypographyLink } from "@/ui/components/typography/TypographyLink";
-import { opn } from "@/util/api";
 
 import { setModRequirementHidden } from "../../actions/persistent";
 import { MOD_REQUIREMENTS_CHECK_ID } from "../../checks/modRequirementsCheck";
-import { EntryActions } from "../../components/entry_actions/EntryActions";
-import { FileRequirement } from "../../components/file_requirement/FileRequirement";
-import { ModRequirement } from "../../components/mod_requirement/ModRequirement";
-import { useModRequirementActions } from "../../components/mod_requirement/useModRequirementActions";
-import { PremiumModal } from "../../components/premium_modal/PremiumModal";
-import { allModRequirements, getModFiles, hiddenModRequirements } from "../../selectors";
+import { DetailView } from "../../components/mod_requirement/DetailView";
+import { ListingRow } from "../../components/mod_requirement/ListingRow";
+import { allModRequirements } from "../../selectors";
 import type { IModRequirementExt } from "../../types";
-import type {
-  IBulkInstallItem,
-  IDetailViewProps,
-  IHealthCheckContent,
-  IListingRowProps,
-} from "./types";
-
-// todo break this into smaller files like the FileRequirementsContent
-// fix the broken styling on the View component
-
-function isModHidden(
-  state: Parameters<typeof hiddenModRequirements>[0],
-  mod: IModRequirementExt,
-): boolean {
-  return (hiddenModRequirements(state)[mod.requiredBy.modId] || []).includes(mod.id);
-}
-
-function ModRequirementsListingRow({
-  api,
-  entry,
-  isHidden,
-  onOpen,
-  onToggleHide,
-}: IListingRowProps) {
-  return (
-    <ModRequirement
-      api={api}
-      isHidden={isHidden}
-      requirementInfo={entry.data as IModRequirementExt}
-      severity={entry.severity}
-      onClick={onOpen}
-      onToggleHide={onToggleHide}
-    />
-  );
-}
-
-/**
- * Detail view for a single missing mod requirement. Mirrors the standalone
- * health-check detail page that shipped on master, in Rich's new card design:
- * install (premium-gated for free users), open mod page, external-requirement
- * confirmation, author notes, thumbs feedback, and hide. Install/feedback logic
- * is shared with the listing row via useModRequirementActions.
- */
-function ModRequirementsDetailView({ entry, api, onBack }: IDetailViewProps) {
-  const { t } = useTranslation(["health_check", "common"]);
-  const mod = entry.data as IModRequirementExt;
-
-  const {
-    givenFeedback,
-    showPremiumAd,
-    showPremiumModal,
-    setShowPremiumModal,
-    openModPage,
-    installInApp,
-    handlePositiveFeedback,
-    handleFeedbackSuccess,
-  } = useModRequirementActions(api, mod, onBack);
-
-  const modFiles = useSelector((state: IState) => getModFiles(state, mod.modId));
-
-  const hiddenRequirementMap = useSelector(hiddenModRequirements);
-  const isHidden = useMemo(
-    () => (hiddenRequirementMap[mod.requiredBy.modId] ?? []).includes(mod.id),
-    [hiddenRequirementMap, mod.requiredBy.modId, mod.id],
-  );
-
-  // Fetch the required mod's files for the preview/version; skipped for external
-  // requirements, which have no Nexus files.
-  useEffect(() => {
-    if (mod.externalRequirement) {
-      return;
-    }
-    getModFilesWithCache(api, mod.gameId, mod.modId).catch((err: unknown) => {
-      log("warn", "health check: failed to fetch requirement mod files", unknownToError(err));
-    });
-  }, [api, mod.gameId, mod.modId, mod.externalRequirement]);
-
-  const openRequiringModPage = useCallback(() => {
-    if (mod.requiredBy.modUrl) {
-      opn(mod.requiredBy.modUrl).catch(() => undefined);
-    }
-  }, [mod.requiredBy.modUrl]);
-
-  const handleToggleHide = useCallback(() => {
-    api.store?.dispatch(setModRequirementHidden(mod.requiredBy.modId, mod.id, !isHidden));
-    onBack();
-  }, [api, mod.requiredBy.modId, mod.id, isHidden, onBack]);
-
-  // External installs can't be auto-detected, so confirming just hides the
-  // requirement from future checks.
-  const handleConfirmInstall = useCallback(() => {
-    api.store?.dispatch(setModRequirementHidden(mod.requiredBy.modId, mod.id, true));
-    onBack();
-  }, [api, mod.requiredBy.modId, mod.id, onBack]);
-
-  const mainFile = modFiles?.[0];
-  const fileData: IFileRequirementData = {
-    fileUID: mod.uid,
-    adultContent: mainFile?.adultContent ?? false,
-    // modName carries the Nexus mod name or, for external requirements, the
-    // external resource name; modUrl is only a last resort.
-    modName: mod.modName || mod.modUrl || "",
-    modDescription: mainFile?.modSummary ?? "",
-    modImageSrc: mainFile?.thumbnailUrl ?? "",
-    fileName: mainFile?.name ?? "",
-    fileVersion: mainFile?.version ?? "",
-    installed: false,
-    enabled: false,
-  };
-
-  const actions = mod.externalRequirement ? (
-    mod.modUrl ? (
-      <Button
-        appearance="moderate"
-        brand="neutral"
-        leftIconPath={mdiOpenInNew}
-        size="sm"
-        onClick={openModPage}
-      >
-        {t("detail::item::open_external_mod_page")}
-      </Button>
-    ) : null
-  ) : (
-    <>
-      {mod.modUrl ? (
-        <Button
-          appearance="moderate"
-          brand="neutral"
-          leftIconPath={mdiOpenInNew}
-          size="sm"
-          onClick={openModPage}
-        >
-          {t("detail::item::install_via_mod_page")}
-        </Button>
-      ) : null}
-
-      <Button
-        appearance="strong"
-        brand="neutral"
-        leftIconPath={mdiDownload}
-        rightIcon={showPremiumAd ? <PremiumBadge /> : undefined}
-        size="sm"
-        onClick={() => void installInApp()}
-      >
-        {t("detail::item::install_one_click")}
-      </Button>
-    </>
-  );
-
-  return (
-    <div className="space-y-6">
-      <div className="flex items-start justify-between gap-x-4">
-        <div className="min-w-0">
-          <Typography className="font-semibold">
-            {t("detail::item::title", { modName: mod.requiredBy.modName })}
-          </Typography>
-
-          <Typography appearance="moderate">
-            <Trans
-              components={{
-                modLink: (
-                  <TypographyLink
-                    brand="primary"
-                    typographyType="inherit"
-                    variant="secondary"
-                    onClick={openRequiringModPage}
-                  />
-                ),
-              }}
-              i18nKey="detail::item::description"
-              ns="health_check"
-              values={{ modName: mod.requiredBy.modName }}
-            />
-          </Typography>
-        </div>
-
-        <EntryActions
-          givenFeedback={givenFeedback}
-          isHidden={isHidden}
-          severity={entry.severity}
-          variant="detail"
-          onHelpful={handlePositiveFeedback}
-          onNotHelpful={handleFeedbackSuccess}
-          onToggleHide={handleToggleHide}
-        />
-      </div>
-
-      {mod.notes ? (
-        <Typography appearance="moderate" typographyType="body-sm">
-          {t("detail::item::author_note", { note: mod.notes })}
-        </Typography>
-      ) : null}
-
-      <FileRequirement actions={actions} file={fileData} />
-
-      {mod.externalRequirement && (
-        <div className="flex items-center gap-x-3 rounded-sm bg-info-weak/20 p-3">
-          <Typography appearance="moderate" as="div" className="grow" typographyType="body-sm">
-            {t("detail::item::after_installing")}
-          </Typography>
-
-          <Button
-            appearance="moderate"
-            brand="neutral"
-            leftIconPath={mdiCheck}
-            size="sm"
-            onClick={handleConfirmInstall}
-          >
-            {t("detail::item::confirm_install")}
-          </Button>
-        </div>
-      )}
-
-      <PremiumModal
-        isOpen={showPremiumModal}
-        onClose={() => setShowPremiumModal(false)}
-        onDownload={() => {
-          setShowPremiumModal(false);
-          openModPage();
-        }}
-      />
-    </div>
-  );
-}
+import { isModHidden } from "./modRequirementEntries";
+import type { IBulkInstallItem, IHealthCheckContent } from "./types";
 
 export const modRequirementsContent: IHealthCheckContent = {
   selectEntries: (state) =>
@@ -258,8 +19,8 @@ export const modRequirementsContent: IHealthCheckContent = {
       severity: "suggestion",
       data: mod,
     })),
-  ListingRow: ModRequirementsListingRow,
-  DetailView: ModRequirementsDetailView,
+  ListingRow,
+  DetailView,
   supportsHide: true,
   isHidden: (state, entry) => isModHidden(state, entry.data as IModRequirementExt),
   toggleHide: (api, entry) => {
